Redesign the whole parser infrastructure to use a stack of alternative continuations, rather than passing individual success and failure continuations to every parser.
This should substantially improve space safety, and it substantially
simplifies a good chunk of the code: no longer is there a confusing
CONSUMED? parameter passed every which way to the continuations.
Several of the parsers were changed to be space-safer independent of
the new stack management, too.  (The code still relies on flat
environments, however, and will lose miserably in the presence of
nested ones.  Perhaps I shall fix that some other day, when the
weather is less painfully stiflingly hot and humid.)

I also took the liberty of restructuring some of the parser's
interface, so that PARSE now accepts a success continuation as well
as a failure continuation, and passes to both the final stream.