Rework *PARSER's syntax, DEFINE-PARSER's semantics, and the repeated parser interface.
*PARSER's syntax has been changed to

  (*PARSER <command> ... <tail>),

with no option for a loser; use PARSER:ON-FAILURE manually.

The delaying of parsers has been moved to DEFINE-PARSER from *PARSER;
*PARSER is now much simpler, and DEFINE-PARSER can now work for
examples like

  (DEFINE-PARSER FOO-PARSER:BAR
    (PARSER:CHOICE FOO-PARSER:BAZ
                   FOO-PARSER:QUUX))

  (DEFINE-PARSER FOO-PARSER:QUUX ...),

where FOO-PARSER:QUUX is named in the lexically prior to its
definition.

The repeating parsers now have parameters for the method of
accumulating results, instantiated for ignoring the terms and for
building lists of the terms.